Added library code and script to run power analysis simulations.
This simulates mating events between mothers and real candidate father
genotypes, and generates families of full siblings of the same sizes as in
the real data set. This then runs FAPS functions on those data and returns
probabilities of paternity for true sires based on paternity alone, after
clustering into sibships, and after clustering including information about
dispersal.

It also returns the posterrior probs that true fathers were missing (1) when
they are really present and (2) when they are truly missing.

 On branch main
 Changes to be committed:
	new file:   02_library/python/amajusmating/simulations.py
	new file:   03_analysis/05_simulations/simulations.py