Plot mating events
Added analysis and plots for the relationship between array size and number of
sibships, and used this to infer the number of missing fathers.

The latter is based on a GLM using mating events averaged over iterations. There
is some funkiness with mating events with <2 offspring which I think I spurious.
This might be worth repeating using a separate GLM for each iteration.

 On branch main
 Changes to be committed:
	new file:   03_analysis/04_mating_events/n_mating_events.R
	new file:   04_figures/02_mating_events/plot_mating_events.R