Added assortative mating back in
12th November

Added lines in the dictionaries for inital parameters and proposal sigmas to
include assortment in the MCMC analyses (most of the code to do that was there
already, lurking behind if statements).

I also added three lines to `check_parameters()` and `run_MCMC()` to include
an estimate of the 'null assortmemt', or the probability of mating between
plants of the same type under random mating, given current dispersal parameters.

 On branch assortment
 Changes to be committed:
	modified:   02_library/python/amajusmating/mcmc.py
	modified:   02_library/python/setup.py
	modified:   03_analysis/03_mcmc/01_mcmc_restrict_kurtosis/output/chain1.out
	modified:   03_analysis/03_mcmc/01_mcmc_restrict_kurtosis/run_MCMC.py
	modified:   03_analysis/03_mcmc/02_mcmc_short_range_kurtosis/run_MCMC.py
	modified:   03_analysis/03_mcmc/03_mcmc_strong_prior_on_kurtosis/run_MCMC.py
	modified:   03_analysis/03_mcmc/04_mcmc_022_missing_fathers/run_MCMC.py
	modified:   03_analysis/03_mcmc/05_mcmc_042_missing_fathers/run_MCMC.py
	modified:   README.md
	modified:   changelog