Set hard prior on q to 0.5
30th May 2023

Set the prior on the proportion of missing fathers to a hard prior of 0.5.
This sets equal prior weight on probability mass from sampled and unsampled
candidates, meaning that the signal all comes from the data. I also set this up
to run as a parallel batch job.

This is because when q is allowed to vary the model moves straight to q=1 -
everyone is compatibly with a missing father, so this solution has high
likelihood. I also don't like running separate analyses for different values of
q because this becomes very cumbersome to run and present, and doesn't really
change much. Instead, I'll discuss other ways to establish who the proportion of
missing fathers.

This is initially done for the 'main' MCMC to check it works. Still need to
repeat this for a model with no mixture, and possibly include assortment.

 Please enter the commit message for your changes. Lines starting
 On branch main
 Changes to be committed:
	modified:   03_analysis/01_data_formatting/setup_FAPS_GPS.py
	new file:   03_analysis/03_mcmc/01_mcmc_main/job_submission.sh
	modified:   03_analysis/03_mcmc/01_mcmc_main/run_MCMC.py